List Contents of local project folder

Retrieves names, timestamps, and sizes of files and folders inside local project folder.

Usage

cloud_local_ls(
  path = "",
  root = ".",
  recursive = FALSE,
  full_names = FALSE,
  ignore = TRUE
)

Arguments

path

(optional) Path, relative to the specified root to list contents of. By default, when path = "", lists root-level files and folders.

root

Local directory path relative to which all other paths are considered.

recursive

(logical) If TRUE, lists contents recursively in all nested subfolders. Default is FALSE.

full_names

(logical) If TRUE, folder path is appended to object names to give a relative file path.

ignore

Logical flag indicating whether to ignore certain directories. Currently, if set to TRUE, the 'renv' folder is ignored due to its typically large size. This parameter may be expanded in the future to support more complex ignore patterns.

Value

A tibble containing the names, last modification timestamps, and sizes in bytes of files and folders inside the specified local folder.

Examples

# list only root-level files and folders
cloud_local_ls() 
#> # A tibble: 27 × 4
#>    name                    type  last_modified       size_b
#>    <chr>                   <chr> <dttm>               <int>
#>  1 Rplot001.png            png   2023-10-18 18:08:02   1011
#>  2 check_args.html         html  2023-10-18 18:07:52   7790
#>  3 check_bool.html         html  2023-10-18 18:07:52   8995
#>  4 check_class.html        html  2023-10-18 18:07:52   8911
#>  5 check_length.html       html  2023-10-18 18:07:52   8894
#>  6 check_null_cond.html    html  2023-10-18 18:07:53   8513
#>  7 check_scalar.html       html  2023-10-18 18:07:53   8355
#>  8 cli_yeah.html           html  2023-10-18 18:07:53   9786
#>  9 cloud_drive_attach.html html  2023-10-18 18:07:54  10012
#> 10 cloud_drive_browse.html html  2023-10-18 18:07:54  10180
#> # ℹ 17 more rows

# list all files in all nested folders
cloud_local_ls(recursive = TRUE)
#> # A tibble: 29 × 4
#>    name                    type  last_modified       size_b
#>    <chr>                   <chr> <dttm>               <int>
#>  1 Rplot001.png            png   2023-10-18 18:08:02   1011
#>  2 check_args.html         html  2023-10-18 18:07:52   7790
#>  3 check_bool.html         html  2023-10-18 18:07:52   8995
#>  4 check_class.html        html  2023-10-18 18:07:52   8911
#>  5 check_length.html       html  2023-10-18 18:07:52   8894
#>  6 check_null_cond.html    html  2023-10-18 18:07:53   8513
#>  7 check_scalar.html       html  2023-10-18 18:07:53   8355
#>  8 cli_yeah.html           html  2023-10-18 18:07:53   9786
#>  9 cloud_drive_attach.html html  2023-10-18 18:07:54  10012
#> 10 cloud_drive_browse.html html  2023-10-18 18:07:54  10180
#> # ℹ 19 more rows

if (FALSE) {
# list contents of "plots/barplots" subfolder (if it exists)
cloud_local_ls("plots/barplots")
}