Write an object to S3

Writes an R object to a file in project's S3 folder. In most cases it is expected to be used for writing data.frames as csv, sav or any other tabular data formats to S3. But it is also possible to save any R object if a proper saving function is provided. It tries to guess a suitable writing function and the output file name where possible.

Usage

cloud_s3_write(x, file, fun = NULL, ..., local = FALSE, root = NULL)

Arguments

x

an R object (e.g. data frame) to write to S3.

file

Path to a file relative to project folder root. Can contain only letters, digits, '-', '_', '.', spaces and '/' symbols.

fun

a function to write a file to cloud location to which x and a file path will be passed (in that order). By default, if fun = NULL, it will be attempted to find an appropriate writing function based on the file extension.

...

further parameters to pass to fun

local

(logical) If TRUE, will additionally create a local file at the corresponding path. Default is FALSE.

root

S3 path of the project root -- point relative to which to consider all relative paths. When left as NULL, the root is automatically derived from the cloudfs.s3 field of the project's DESCRIPTION file.

Default writing functions

Here's how we identify a writing function based on file extension

.csv: readr::write_csv

.json: jsonlite::write_json

.rds: base::saveRDS

.xls: writexl::write_xlsx

.xlsx: writexl::write_xlsx

.sav: haven::write_sav

.xml: xml2::write_xml

Examples

if (FALSE) {
# write mtcars dataframe to mtcars.csv in data folder
cloud_s3_write(mtcars, "data/mtcars.csv")
cloud_s3_write(random_forest, "models/random_forest.rds")

# provide custom writing function with parameters 
cloud_s3_write(c("one", "two"), "text/count.txt", writeLines, sep = "\n\n")
}